\name{AlonDS}

\docType{data}

\alias{AlonDS}

\title{Alon Colon Cancer Data Set}

\description{This data set was collected by Alon \emph{et. al.} and consists of 2000 genes measured on 62 patients: 40 diagnosed with colon cancer and 22 healthy patients. The patient status is described by the factor \sQuote{grouping} and the gene values are given by the numeric variables
\sQuote{genes.1} through \sQuote{genes.2000}.}

\usage{data(AlonDS)}

\references{ Alon, U., Barkai, N., Notterman, D.A., Gish, K., Ybarra, S., Mack, D. and Levine, A.J. (1999) \dQuote{Broad patterns of gene expression revealed by clustering analysis of tumor and normal colon tissues probed by oligonucleotide arrays}, In: \emph{Proceedings National Academy of Sciences} USA 96, 6745-6750.}

\format{A data frame containing 62 observations on one factor, and 2000 numeric variables.}
\keyword{datasets}


