\name{SinghDS}
\docType{data}
\alias{SinghDS}
\title{Singh Prostate Cancer Data Set}
\description{This data set was collected by Singh \emph{et. al.} and consists of 6033 genes measured on 102 patients, 52 diagnosed with prostate cancer and 50 healthy patients. The patient status is described by the factor \sQuote{grouping} and the gene values are given by the numeric variables
\sQuote{genes.1} through \sQuote{genes.6033}.}
\usage{data(SinghDS)}
\references{Singh, D., Febbo, P. G., Ross, K., Jackson, D. G., Manola, J., Ladd, C., Tamayo, P., Renshaw, A. A., D' Amico, A. V., Richie, J. P., Lander, E. S., Loda, M., Kantoff, P. W., Golub, T. R. and Sellers, W. R. (2002). Gene expression correlates of clinical prostate cancer behavior. \emph{Cancer Cell}, 1 (2), 203-209.}
\format{A data frame containing 102 observations on one factor, and 6033 numeric variables.}
\keyword{datasets}


