\name{HiDimDA-internal}

\alias{RepLOptim}
\alias{buildB}
\alias{f}
\alias{f1}
\alias{fgrad}
\alias{fgrad1}
\alias{fhess}
\alias{fhess1}
\alias{grpmeans}
\alias{grpvar}
\alias{l2vnorm}
\alias{scalebygrps}
\alias{*.SigFq}
\alias{/.SigFq}
\alias{*.SigFqInv}
\alias{/.SigFqInv}
\alias{-.SigFq}
\alias{-.SigFqInv}
\alias{+.SigFq}
\alias{+.SigFqInv}
\alias{EvalClrule}
\alias{mylocmle}
\alias{mylocfdr}
\alias{loccov}
\alias{loccov2}
\alias{locfdrpval}
\alias{HC}
\alias{tscores}
\alias{fscores}

\title{Internal HiDimDA Functions}
\description{
  Internal HiDimDA functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
